% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusData.R
\name{aggregateByAgeGroups}
\alias{aggregateByAgeGroups}
\title{Aggregate population by age groups}
\usage{
aggregateByAgeGroups(ages, pops, age_groups)
}
\description{
Aggregate population by age groups
}
\keyword{internal}
