% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMatrixPolymod.R
\name{contactMatrixPolymod}
\alias{contactMatrixPolymod}
\title{Calculate a contact matrix for age groups based on Polymod contact survey data}
\usage{
contactMatrixPolymod(agelims, agepops = NULL)
}
\arguments{
\item{agelims}{minimum age in years for each age group}

\item{agepops}{population size of each group, defaulting to demography of Polymod survey population}
}
\description{
Calculate a contact matrix for age groups based on Polymod contact survey data
}
