% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMatrixPropPref.R
\name{contactMatrixPropPref}
\alias{contactMatrixPropPref}
\title{Calculate group contact matrix with proportional mixing and preferential mixing within group}
\usage{
contactMatrixPropPref(popsize, contactrate, ingroup)
}
\arguments{
\item{popsize}{population size of each group}

\item{contactrate}{overall contact rate of each group}

\item{ingroup}{fraction of each group's contacts that are exclusively in-group}
}
\description{
Calculate group contact matrix with proportional mixing and preferential mixing within group
}
