% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSize.R
\name{getFinalSize}
\alias{getFinalSize}
\title{Calculate final size of outbreak: the total number of infections in each group}
\usage{
getFinalSize(
  vacTime,
  vacPortion,
  popSize,
  R0,
  recoveryRate,
  relContact,
  contactWithinGroup,
  relSusc
)
}
\arguments{
\item{vacTime}{time after first case at which all vaccinations are delivered}

\item{vacPortion}{fraction of each population vaccinated}

\item{popSize}{size of each population}

\item{R0}{overall basic reproduction number}

\item{recoveryRate}{inverse of mean infectious period
(same time units as vacTime)}

\item{relContact}{relative overall contact rate of each group}

\item{contactWithinGroup}{fraction of each group's contacts that are
in-group vs out-group}

\item{relSusc}{relative susceptibility to infection per contact of each group}
}
\description{
Calculate final size of outbreak: the total number of infections in each group
}
