% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSizeDist.R
\name{getFinalSizeDist}
\alias{getFinalSizeDist}
\title{Estimate the distribution of final outbreak sizes by group using stochastic simulations of multi-group model}
\usage{
getFinalSizeDist(n, transmrates, recoveryrate, popsize, initR, initI, initV)
}
\arguments{
\item{n}{the number of simulations to run}

\item{transmrates}{matrix of group-to-group (column-to-row) transmission rates}

\item{recoveryrate}{inverse of mean infectious period}

\item{popsize}{the population size of each group}

\item{initR}{initial number of each group already infected and removed for each simulation (included in final size result); if a matrix, row i is used for simulation i}

\item{initI}{initial number of each group infectious for each simulation; if a matrix, row i is used for simulation i}

\item{initV}{initial number of each group immune due to vaccination or prior-outbreak infection for each simulation (not included in final size result); if a matrix, row i is used for simulation i}
}
\value{
a matrix with the final number infected from each group (column) in each simulation (row)
}
\description{
Estimate the distribution of final outbreak sizes by group using stochastic simulations of multi-group model
}
\examples{
getFinalSizeDist(n = 10, transmrates = matrix(0.2, 2 ,2), recoveryrate = 0.3,
popsize = c(100, 150), initR = c(0, 0), initI = c(0, 1), initV = c(10, 10))
}
