% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullHistogram.R
\name{getFullHistogram}
\alias{getFullHistogram}
\title{Gets full histogram of final sizes for an age-structured population}
\usage{
getFullHistogram(R0, agelims, agepops, agecovr, ageveff, initgrp)
}
\arguments{
\item{R0}{basic reproduction number}

\item{agelims}{Vector of age group limits}

\item{agepops}{Vector of population sizes for each age group}

\item{agecovr}{Vector of vaccination coverage rates for each age group}

\item{ageveff}{Vector of vaccine effectiveness rates for each age group}

\item{initgrp}{Index of the age group where the initial infection occurs}
}
\value{
A data frame with columns for final size and frequency
}
\description{
Gets full histogram of final sizes for an age-structured population
}
