% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusData.R
\name{processCensusDataBySex}
\alias{processCensusDataBySex}
\title{Process census data with sex disaggregation}
\usage{
processCensusDataBySex(county_data, age_groups)
}
\description{
Process census data with sex disaggregation
}
\keyword{internal}
