% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simodeSIR.R
\name{simodeSIR}
\alias{simodeSIR}
\title{Calculate time trajectory of solution to a Susceptible-Infectious-Removed (SIR) system by
numerically solving the multi-group ordinary differential equations}
\usage{
simodeSIR(transmrates, recoveryrate, popsize, initR, initI, initV, times = NA)
}
\arguments{
\item{transmrates}{matrix of group-to-group (column-to-row) transmission rates}

\item{recoveryrate}{inverse of mean infectious period}

\item{popsize}{the population size of each group}

\item{initR}{initial number of each group already infected and removed (included in final size)}

\item{initI}{initial number of each group infectious}

\item{initV}{initial number of each group vaccinated}

\item{times}{optional argument for times at which the solution is to be calculated; if NA (the
default) the solution will be computed up to the time at which the total number of infectious
individuals is sufficiently close to zero (outbreak extinction)}
}
\value{
vector of final sizes (number of infected over whole outbreak) for each group
}
\description{
Calculate time trajectory of solution to a Susceptible-Infectious-Removed (SIR) system by
numerically solving the multi-group ordinary differential equations
}
\examples{
simodeSIR(transmrates = matrix(0.2, 2 ,2), recoveryrate = 0.3,
popsize = c(100, 150), initR = c(0, 0), initI = c(0, 1), initV = c(10, 10))
}
