% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.equity-package.R
\docType{package}
\name{vaccine.equity-package}
\alias{vaccine.equity}
\alias{vaccine.equity-package}
\title{vaccine.equity: What the Package Does (Title Case)}
\description{
The purpose of this package is
}
\keyword{internal}
