% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMatrixAgeSchool.R
\name{contactMatrixAgeSchool}
\alias{contactMatrixAgeSchool}
\title{Calculate a contact matrix for age groups and schools}
\usage{
contactMatrixAgeSchool(
  agelims,
  agepops,
  schoolagegroups,
  schoolpops,
  schportion
)
}
\arguments{
\item{agelims}{minimum age in years for each age group}

\item{agepops}{population size of each age group}

\item{schoolagegroups}{index of the age group covered by each school}

\item{schoolpops}{population size of each school}

\item{schportion}{portion of within-age-group contacts that are exclusively within school}
}
\description{
Calculate a contact matrix for age groups and schools
}
