% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactMatrixPropPref.R
\name{contactMatrixPropPref}
\alias{contactMatrixPropPref}
\title{Calculate group contact matrix with proportional mixing and preferential mixing within group}
\usage{
contactMatrixPropPref(popsize, contactrate, ingroup)
}
\arguments{
\item{popsize}{population size of each group}

\item{contactrate}{overall contact rate of each group}

\item{ingroup}{fraction of each group's contacts that are exclusively in-group}
}
\value{
a square matrix with the contact rate of each group (row) with members of each
other group (column)
}
\description{
Calculate group contact matrix with proportional mixing and preferential mixing within group
}
\examples{
contactMatrixPropPref(popsize = c(100, 150, 200), contactrate = c(1.1, 1, 0.9),
ingroup = c(0.2, 0.25, 0.22))
}
