% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSizeAnalytic.R
\name{getFinalSizeAnalytic}
\alias{getFinalSizeAnalytic}
\title{Calculate final size of outbreak: the total number of infections in each group,
by solving the analytic final size equation}
\usage{
getFinalSizeAnalytic(transmrates, recoveryrate, popsize, initR, initI, initV)
}
\arguments{
\item{transmrates}{matrix of group-to-group (column-to-row) transmission rates}

\item{recoveryrate}{inverse of mean infectious period}

\item{popsize}{the population size of each group}

\item{initR}{initial number of each group already infected and removed (included in final size)}

\item{initI}{initial number of each group infectious}

\item{initV}{initial number of each group vaccinated}
}
\description{
Calculate final size of outbreak: the total number of infections in each group,
by solving the analytic final size equation
}
