% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSizeAnalytic.R
\name{getFinalSizeAnalytic}
\alias{getFinalSizeAnalytic}
\title{TODO: Add documentation and @internal flag unless we want it to be exported}
\usage{
getFinalSizeAnalytic(Rinit, Iinit, Vinit, N, R0, a, eps, q)
}
\description{
TODO: Add documentation and @internal flag unless we want it to be exported
}
