% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSizeDist.R
\name{getFinalSizeDist}
\alias{getFinalSizeDist}
\title{Estimate the distribution of final outbreak sizes by group using stochastic simulations of multi-group model}
\usage{
getFinalSizeDist(n, transmrates, recoveryrate, popsize, initR, initI, initV)
}
\arguments{
\item{n}{the number of simulations to run}

\item{transmrates}{matrix of group-to-group (column-to-row) transmission rates}

\item{recoveryrate}{inverse of mean infectious period}

\item{popsize}{the population size of each group}

\item{initR}{initial number of each group already infected and removed (included in size result)}

\item{initI}{initial number of each group infectious}

\item{initV}{initial number of each group vaccinated}
}
\value{
a matrix with the final number infected from each group (column) in each simulation (row)
}
\description{
Estimate the distribution of final outbreak sizes by group using stochastic simulations of multi-group model
}
