% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalSizeDistEscape.R
\name{getFinalSizeDistEscape}
\alias{getFinalSizeDistEscape}
\title{TODO: Add documentation and @internal flag unless we want it to be exported}
\usage{
getFinalSizeDistEscape(
  n,
  transmrates,
  recoveryrate,
  popsize,
  initR,
  initI,
  initV
)
}
\description{
TODO: Add documentation and @internal flag unless we want it to be exported
}
