% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSizeAtTime.R
\name{getSizeAtTime}
\alias{getSizeAtTime}
\title{Calculate outbreak size at a given time}
\usage{
getSizeAtTime(
  time,
  R0,
  recoveryRate,
  popsize,
  initR,
  initI,
  initV,
  contactmatrix,
  relcontact,
  relsusc
)
}
\arguments{
\item{time}{the time at which to calculate the outbreak size}

\item{R0}{overall basic reproduction number}

\item{recoveryRate}{inverse of the mean duration of infectious period}

\item{popsize}{population size of each group}

\item{initR}{initial number of each group who have already recovered from infection}

\item{initI}{initial number of each group who are actively infected}

\item{initV}{initial number of each group who are vaccinated and had no prior infection}

\item{contactmatrix}{matrix with fraction of each row-group's contacts that are with column-group}

\item{relcontact}{relative overall contact rates of each group}

\item{relsusc}{relative susceptibility to infection per contact of each group}
}
\value{
a list with totalSize (total cumulative infections) and activeSize (total currently infected) in each group at the specified time
}
\description{
Calculate outbreak size at a given time
}
