% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusData.R
\name{listCounties}
\alias{listCounties}
\title{List available counties for a state}
\usage{
listCounties(state_fips, year = 2024)
}
\arguments{
\item{state_fips}{Two-digit FIPS code for the state}

\item{year}{Census year (2020-2024), default 2024}
}
\value{
Character vector of county names
}
\description{
List available counties for a state
}
\examples{
\dontrun{
utah_counties <- listCounties(state_fips = "49", year = 2024)
}
}
