% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeSIR.R
\name{odeSIR}
\alias{odeSIR}
\title{Ordinary differential equation function for multi-group susceptible-infectious-removed (SIR) model
used as "func" argument passed to the ode() function from deSolve package}
\usage{
odeSIR(time, state, par)
}
\arguments{
\item{time}{vector of times at which the function will be evaluated}

\item{state}{vector of number of individuals in each group at each state: S states
followed by I states followed by R states}

\item{par}{vector of parameter values: group-to-group transmission rate matrix elements (row-wise)
followed by recovery rate}
}
\description{
Ordinary differential equation function for multi-group susceptible-infectious-removed (SIR) model
used as "func" argument passed to the ode() function from deSolve package
}
