% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeSIR.R
\name{odeSIR}
\alias{odeSIR}
\title{ODE system for multi-group SIR model
TODO: Add documentation and @keywords internal flag unless we want it to be exported
time parameter doesn't seem to be used, can this be removed safely?}
\usage{
odeSIR(time, state, par)
}
\description{
ODE system for multi-group SIR model
TODO: Add documentation and @keywords internal flag unless we want it to be exported
time parameter doesn't seem to be used, can this be removed safely?
}
\keyword{internal}
