% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissionRates.R
\name{transmissionRates}
\alias{transmissionRates}
\title{Calculate transmission rate matrix for multigroup model with specified R0}
\usage{
transmissionRates(R0, meaninf, popsize, incontact, relcontact, relsusc)
}
\arguments{
\item{R0}{overall basic reproduction number}

\item{meaninf}{mean duration of infectious period}

\item{popsize}{population size of each group}

\item{incontact}{fraction of each group's contacts that are exclusively within group}

\item{relcontact}{relative overall contact rates of each group}

\item{relsusc}{relative susceptibility to infection per contact of each group}
}
\description{
Calculate transmission rate matrix for multigroup model with specified R0
}
